:bug: get proper container name when using weird branch names
