+++ apps/owidbot/chart_diff.py
@@ -4,7 +4,8 @@ from structlog import get_logger
 
 from apps.chart_sync.cli import _modified_chart_ids_by_admin
 from apps.wizard.pages.chart_diff.chart_diff import ChartDiffModified
-from apps.wizard.utils.env import OWID_ENV, OWIDEnv, get_container_name
+from apps.wizard.utils.env import OWID_ENV, OWIDEnv
+from etl.config import get_container_name
 
 from . import github_utils as gh_utils
