+++ tests/apps/wizard/utils/test_env.py
@@ -1,10 +1,4 @@
-from apps.wizard.utils.env import Config, OWIDEnv, get_container_name
-
-
-def test_get_container_name():
-    assert get_container_name("branch") == "staging-site-branch"
-    assert get_container_name("feature/x") == "staging-site-feature-x"
-    assert get_container_name("do_not-do/this") == "staging-site-do-not-do-this"
+from apps.wizard.utils.env import Config, OWIDEnv
 
 
 def test_OWIDEnv_staging():
